(feat) edit form

+++ app/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, DecimalField
+from wtforms import StringField, DecimalField, IntegerField
 from wtforms.validators import DataRequired
 
 class ProductForm(FlaskForm):
@@ -9,5 +9,5 @@ class ProductForm(FlaskForm):
 
 class ItemForm(FlaskForm):
     name = StringField('name', validators=[DataRequired()])
-    available_qty = DecimalField('available_qty', validators=[DataRequired()])
+    available_qty = IntegerField('available_qty', validators=[DataRequired()])
